:sparkles: #61 메모좋아요 취소 기능 구현

+++ src/main/java/com/seollem/server/memolike/MemoLike.java
@@ -23,7 +23,7 @@ public class MemoLike extends Auditable {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int memoLikeId;
+  private long memoLikeId;
 
   @ManyToOne
   @JoinColumn(name = "MEMO_ID")
